cloudrun/agent_charts_signal/requirements.txt
-e ../../shared
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2
google-cloud-storage>=2.14
python-dotenv>=1.0
anthropic>=0.34
openai>=1.0
httpx>=0.27
